+++ zara/plugins/__init__.py
@@ -9,6 +9,7 @@ from .api import (
     PluginRuntime,
     RuntimeStatus,
     ServicePlugin,
+    StartupUnavailable,
 )
 from .manager import PluginDiagnostic, PluginManager, PluginState
 
@@ -22,6 +23,7 @@ __all__ = [
     "PluginState",
     "RuntimeStatus",
     "ServicePlugin",
+    "StartupUnavailable",
     "ToolCancellation",
     "current_tool_cancellation",
 ]